operacat/requirements.txt
beautifulsoup4<4.6.1
click
Django
django-comments-xtd
django-contrib-comments
django-markup
django-modelcluster
django-registration
django-registration-redux
django-taggit<1.0
django-treebeard
djangorestframework
docutils
Flask
html5lib
itsdangerous
Jinja2
Markdown
MarkupSafe
olefile
Pillow<6.0
psycopg2
pytz
requests
six
Unidecode
urllib3<1.25
wagtail
webencodings
Werkzeug
Willow
